replace VAlidation by InvertableValidation add not and xor to Validator
