+++ src/Validation.spec.ts
@@ -1,313 +1,825 @@
 import {
-    validValidation,
-    invalidValidation,
-    asyncValidation,
+    createValidValidation,
+    createInvalidValidation,
+    createAsyncValidation,
 } from "./Validation";
 
 describe("Validation", () => {
     describe("and", () => {
-        it("Valid and Valid should keep the first Valid value", () => {
-            const res = validValidation().and(validValidation());
+        it("ValidValidation and ValidValidation return a ValidValidation", () => {
+            const date = new Date();
+            const res = createValidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
+            ]).and(
+                createValidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+            );
 
             expect(res.isValid).toBe(true);
             expect(res.getResult()).toBeUndefined();
-        });
-
-        it("Valid and Invalid should keep the Invalid value", () => {
-            const res = validValidation().and(
-                invalidValidation([
-                    { message: "invalid", value: "invalid value" },
-                ])
-            );
-            expect(res.isValid).toBe(false);
-            expect(res.getResult()).toEqual([
-                { message: "invalid", value: "invalid value" },
+            expect(res.results).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: true,
+                    inverted: false,
+                },
             ]);
         });
 
-        it("Valid and asyncValidation(Valid) should return an asyncValidation(Valid)", async () => {
-            const res = validValidation().and(
-                asyncValidation.validValidation()
-            );
-            expect(res.fork).toBeDefined();
-
-            expect(await res.getResult()).toBe(undefined);
-
-            expect(await res.getResult()).toBeUndefined();
-        });
-
-        it("Valid and asyncValidation(InValid) should return an asyncValidation(Invalid)", async () => {
-            const res = validValidation().and(
-                asyncValidation.invalidValidation([
-                    { message: "invalid", value: "invalid value" },
-                ])
-            );
-            expect(res.fork).toBeDefined();
-
-            expect(await res.getResult()).toEqual([
-                { message: "invalid", value: "invalid value" },
-            ]);
-        });
-
-        it("InValid and Valid should keep the Invalid value", () => {
-            const res = invalidValidation([
-                { message: "invalid", value: "invalid value" },
-            ]).and(validValidation());
-            expect(res.isValid).toBe(false);
-            expect(res.getResult()).toEqual([
-                { message: "invalid", value: "invalid value" },
-            ]);
-        });
-
-        it("Invalid and Invalid should concat the two Invalid value", () => {
-            const res = invalidValidation([
-                { message: "invalid1", value: "invalid value" },
+        it("ValidValidation and InvalidValidation should return InvalidValidation", () => {
+            const res = createValidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
             ]).and(
-                invalidValidation([
-                    { message: "invalid2", value: "invalid value" },
+                createInvalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: false,
+                        inverted: false,
+                    },
                 ])
             );
             expect(res.isValid).toBe(false);
             expect(res.getResult()).toEqual([
-                { message: "invalid1", value: "invalid value" },
-                { message: "invalid2", value: "invalid value" },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
+            ]);
+            expect(res.results).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
             ]);
         });
 
-        it("Invalid and asyncValidation(Valid) should return an asyncValidation(Invalid)", async () => {
-            const res = invalidValidation([
-                { message: "invalid", value: "invalid value" },
-            ]).and(asyncValidation.validValidation());
+        it("ValidValidation and AsyncValidation(ValidValidation) should return an AsyncValidation(ValidValidation)", async () => {
+            const res = createValidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
+            ]).and(
+                createAsyncValidation.validValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+            );
             expect(res.fork).toBeDefined();
 
-            expect(await res.getResult()).toEqual([
-                { message: "invalid", value: "invalid value" },
-            ]);
+            expect(await res.getResult()).toBeUndefined();
         });
 
-        it("Invalid and asyncValidation(InValid) should return an asyncValidation(Invalid)", async () => {
-            const res = invalidValidation([
-                { message: "invalid1", value: "invalid value" },
+        it("Valid and AsyncValidation(InValidValidation) should return an AsyncValidation(InvalidValidation)", async () => {
+            const res = createValidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
             ]).and(
-                asyncValidation.invalidValidation([
-                    { message: "invalid2", value: "invalid value" },
+                createAsyncValidation.invalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: false,
+                        inverted: false,
+                    },
                 ])
             );
             expect(res.fork).toBeDefined();
 
             expect(await res.getResult()).toEqual([
-                { message: "invalid1", value: "invalid value" },
-                { message: "invalid2", value: "invalid value" },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
             ]);
         });
 
-        it("asyncValidation(Valid) and Valid should return asyncValidation(Valid)", async () => {
-            const res = asyncValidation
-                .validValidation()
-                .and(validValidation());
+        it("InvalidValidation and ValidValidation should return InvalidValidation", () => {
+            const res = createInvalidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+            ]).and(
+                createValidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+            );
+            expect(res.isValid).toBe(false);
+            expect(res.getResult()).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+            ]);
+            expect(res.results).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: true,
+                    inverted: false,
+                },
+            ]);
+        });
+
+        it("InvalidValidation and InvalidValidation should return InvalidValidation", () => {
+            const res = createInvalidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+            ]).and(
+                createInvalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: false,
+                        inverted: false,
+                    },
+                ])
+            );
+            expect(res.isValid).toBe(false);
+            expect(res.getResult()).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
+            ]);
+            expect(res.results).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
+            ]);
+        });
+
+        it("InvalidValidation and AsyncValidation(ValidValidation) should return an AsyncValidation(InvalidValidation)", async () => {
+            const res = createInvalidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+            ]).and(
+                createAsyncValidation.validValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+            );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+            ]);
+        });
+
+        it("InvalidValidation and AsyncValidation(InValidValidation) should return an AsyncValidation(InvalidValidation)", async () => {
+            const res = createInvalidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+            ]).and(
+                createAsyncValidation.invalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: false,
+                        inverted: false,
+                    },
+                ])
+            );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
+            ]);
+        });
+
+        it("AsyncValidation(ValidValidation) and ValidValidation should return AsyncValidation(ValidValidation)", async () => {
+            const res = createAsyncValidation
+                .validValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+                .and(
+                    createValidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: true,
+                            inverted: false,
+                        },
+                    ])
+                );
             expect(res.fork).toBeDefined();
 
             expect(await res.getResult()).toBeUndefined();
         });
 
-        it("asyncValidation(Valid) and Invalid should return asyncValidation(Invalid)", async () => {
-            const res = asyncValidation
-                .validValidation()
+        it("AsyncValidation(Valid) and Invalid should return asyncValidation(Invalid)", async () => {
+            const res = createAsyncValidation
+                .validValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
                 .and(
-                    invalidValidation([
-                        { message: "invalid", value: "invalid value" },
+                    createInvalidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: false,
+                            inverted: false,
+                        },
                     ])
                 );
             expect(res.fork).toBeDefined();
 
             expect(await res.getResult()).toEqual([
-                { message: "invalid", value: "invalid value" },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
             ]);
         });
 
-        it("asyncValidation(Valid) and asyncValidation(Valid) should return an asyncValidation(Valid)", async () => {
-            const res = asyncValidation
-                .validValidation()
-                .and(asyncValidation.validValidation());
+        it("AsyncValidation(ValidValidation) and AsyncValidation(ValidValidation) should return an AsyncValidation(ValidValidation)", async () => {
+            const res = createAsyncValidation
+                .validValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+                .and(
+                    createAsyncValidation.validValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: true,
+                            inverted: false,
+                        },
+                    ])
+                );
             expect(res.fork).toBeDefined();
-
-            expect(await res.getResult()).toBeUndefined();
         });
 
-        it("asyncValidation(Valid) and asyncValidation(InValid) should return an asyncValidation(Invalid)", async () => {
-            const res = asyncValidation
-                .validValidation()
+        it("AsyncValidation(ValidValidation) and AsyncValidation(InvalidValidation) should return an AsyncValidation(InvalidValidation)", async () => {
+            const res = createAsyncValidation
+                .validValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
                 .and(
-                    asyncValidation.invalidValidation([
-                        { message: "invalid", value: "invalid value" },
+                    createAsyncValidation.invalidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: false,
+                            inverted: false,
+                        },
                     ])
                 );
             expect(res.fork).toBeDefined();
 
             expect(await res.getResult()).toEqual([
-                { message: "invalid", value: "invalid value" },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
             ]);
         });
 
-        it("asyncValidation(Invalid) and Valid should return asyncValidation(Invalid)", async () => {
-            const res = asyncValidation
+        it("AsyncValidation(InvalidValidation) and ValidValidation should return AsyncValidation(InvalidValidation)", async () => {
+            const res = createAsyncValidation
                 .invalidValidation([
-                    { message: "invalid", value: "invalid value" },
-                ])
-                .and(validValidation());
-            expect(res.fork).toBeDefined();
-
-            expect(await res.getResult()).toEqual([
-                { message: "invalid", value: "invalid value" },
-            ]);
-        });
-
-        it("asyncValidation(Invalid) and Invalid should return asyncValidation(Invalid)", async () => {
-            const res = asyncValidation
-                .invalidValidation([
-                    { message: "invalid1", value: "invalid value" },
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: false,
+                        inverted: false,
+                    },
                 ])
                 .and(
-                    invalidValidation([
-                        { message: "invalid2", value: "invalid value" },
+                    createValidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: true,
+                            inverted: false,
+                        },
                     ])
                 );
             expect(res.fork).toBeDefined();
 
             expect(await res.getResult()).toEqual([
-                { message: "invalid1", value: "invalid value" },
-                { message: "invalid2", value: "invalid value" },
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
             ]);
         });
 
-        it("asyncValidation(Invalid) and asyncValidation(Valid) should return an asyncValidation(Invalid)", async () => {
-            const res = asyncValidation
+        it("AsyncValidation(InvalidValidation) and InvalidValidation should return AsyncValidation(InvalidValidation)", async () => {
+            const res = createAsyncValidation
                 .invalidValidation([
-                    { message: "invalid", value: "invalid value" },
-                ])
-                .and(asyncValidation.validValidation());
-            expect(res.fork).toBeDefined();
-
-            expect(await res.getResult()).toEqual([
-                { message: "invalid", value: "invalid value" },
-            ]);
-        });
-
-        it("asyncValidation(Invalid) and asyncValidation(InValid) should return an asyncValidation(Invalid)", async () => {
-            const res = asyncValidation
-                .invalidValidation([
-                    { message: "invalid1", value: "invalid value" },
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: false,
+                        inverted: false,
+                    },
                 ])
                 .and(
-                    asyncValidation.invalidValidation([
-                        { message: "invalid2", value: "invalid value" },
+                    createInvalidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: false,
+                            inverted: false,
+                        },
                     ])
                 );
             expect(res.fork).toBeDefined();
 
             expect(await res.getResult()).toEqual([
-                { message: "invalid1", value: "invalid value" },
-                { message: "invalid2", value: "invalid value" },
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
+            ]);
+        });
+
+        it("AsyncValidation(InvalidValidation) and AsyncValidation(ValidValidation) should return an AsyncValidation(InvalidValidation)", async () => {
+            const res = createAsyncValidation
+                .invalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: false,
+                        inverted: false,
+                    },
+                ])
+                .and(
+                    createAsyncValidation.validValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: true,
+                            inverted: false,
+                        },
+                    ])
+                );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+            ]);
+        });
+
+        it("AsyncValidation(InvalidValidation) and AsyncValidation(InValidValidation) should return an AsyncValidation(InvalidValidation)", async () => {
+            const res = createAsyncValidation
+                .invalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: false,
+                        inverted: false,
+                    },
+                ])
+                .and(
+                    createAsyncValidation.invalidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: false,
+                            inverted: false,
+                        },
+                    ])
+                );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
             ]);
         });
     });
 
     describe("or", () => {
-        it("Valid or Valid should keep the first Valid value", () => {
-            const res = validValidation().or(validValidation());
-            expect(res.isValid).toBe(true);
-            expect(res.getResult()).toBeUndefined();
-        });
-
-        it("Valid or InValid should keep the Valid value", () => {
-            const res = validValidation().or(
-                invalidValidation([
-                    { message: "invalid", value: "invalid value" },
-                ])
-            );
-            expect(res.isValid).toBe(true);
-            expect(res.getResult()).toBeUndefined();
-        });
-
-        it("Valid or asyncValidation(Valid) return asyncValidation(Valid)", async () => {
-            const res = validValidation().or(asyncValidation.validValidation());
-            expect(res.fork).toBeDefined();
-
-            expect(await res.getResult()).toBeUndefined();
-        });
-
-        it("Valid or asyncValidation(InValid) should return asyncValidation(Valid)", async () => {
-            const res = validValidation().or(
-                asyncValidation.invalidValidation([
-                    { message: "invalid", value: "invalid value" },
-                ])
-            );
-            expect(res.fork).toBeDefined();
-
-            expect(await res.getResult()).toBeUndefined();
-        });
-
-        it("Invalid or Valid should keep the Valid value", () => {
-            const res = invalidValidation([
-                { message: "invalid", value: "invalid value" },
-            ]).or(validValidation());
-            expect(res.isValid).toBe(true);
-            expect(res.getResult()).toBeUndefined();
-        });
-
-        it("Invalid or Invalid should concat the Invalid value", () => {
-            const res = invalidValidation([
-                { message: "invalid1", value: "invalid value" },
+        it("ValidValidation or ValidValidation should keep the Valid value", () => {
+            const res = createValidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
             ]).or(
-                invalidValidation([
-                    { message: "invalid2", value: "invalid value" },
+                createValidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+            );
+            expect(res.isValid).toBe(true);
+            expect(res.getResult()).toBeUndefined();
+            expect(res.results).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: true,
+                    inverted: false,
+                },
+            ]);
+        });
+
+        it("ValidValidation or InValidValidation should keep the Valid value", () => {
+            const res = createValidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
+            ]).or(
+                createInvalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: false,
+                        inverted: false,
+                    },
+                ])
+            );
+            expect(res.isValid).toBe(true);
+            expect(res.getResult()).toBeUndefined();
+            expect(res.results).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
+            ]);
+        });
+
+        it("ValidValidation or AsyncValidation(ValidValidation) should return an AsyncValidation(ValidValidation)", async () => {
+            const res = createValidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
+            ]).or(
+                createAsyncValidation.validValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+            );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toBeUndefined();
+        });
+
+        it("Valid or AsyncValidation(InValidValidation) should return an AsyncValidation(ValidValidation)", async () => {
+            const res = createValidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
+            ]).or(
+                createAsyncValidation.invalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: false,
+                        inverted: false,
+                    },
+                ])
+            );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toBeUndefined();
+        });
+
+        it("InvalidValidation or ValidValidation should return ValidValidation", () => {
+            const res = createInvalidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+            ]).or(
+                createValidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+            );
+            expect(res.isValid).toBe(true);
+            expect(res.getResult()).toBeUndefined();
+            expect(res.results).toEqual([
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: true,
+                    inverted: false,
+                },
+            ]);
+        });
+
+        it("InvalidValidation or InvalidValidation should return InvalidValidation", () => {
+            const res = createInvalidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+            ]).or(
+                createInvalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: false,
+                        inverted: false,
+                    },
                 ])
             );
             expect(res.isValid).toBe(false);
             expect(res.getResult()).toEqual([
-                { message: "invalid1", value: "invalid value" },
-                { message: "invalid2", value: "invalid value" },
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
+            ]);
+            expect(res.results).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
             ]);
         });
 
-        it("Invalid or asyncValidation(Valid) return asyncValidation(Valid)", async () => {
-            const res = invalidValidation([
-                { message: "invalid", value: "invalid value" },
-            ]).or(asyncValidation.validValidation());
+        it("InvalidValidation or AsyncValidation(ValidValidation) should return an AsyncValidation(ValidValidation)", async () => {
+            const res = createInvalidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+            ]).or(
+                createAsyncValidation.validValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+            );
             expect(res.fork).toBeDefined();
 
             expect(await res.getResult()).toBeUndefined();
         });
 
-        it("Invalid or asyncValidation(InValid) should return asyncValidation(Valid)", async () => {
-            const res = invalidValidation([
-                { message: "invalid1", value: "invalid value" },
+        it("InvalidValidation or AsyncValidation(InValidValidation) should return an AsyncValidation(InvalidValidation)", async () => {
+            const res = createInvalidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
             ]).or(
-                asyncValidation.invalidValidation([
-                    { message: "invalid2", value: "invalid value" },
+                createAsyncValidation.invalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: false,
+                        inverted: false,
+                    },
                 ])
             );
             expect(res.fork).toBeDefined();
 
             expect(await res.getResult()).toEqual([
-                { message: "invalid1", value: "invalid value" },
-                { message: "invalid2", value: "invalid value" },
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
             ]);
         });
 
-        it("asyncValidation(Valid) or Valid should return asyncValidation(Valid)", async () => {
-            const res = asyncValidation.validValidation().or(validValidation());
-            expect(res.fork).toBeDefined();
-
-            expect(await res.getResult()).toBeUndefined();
-        });
-
-        it("asyncValidation(Valid) or InValid should Return asyncValidation(Valid)", async () => {
-            const res = asyncValidation
-                .validValidation()
+        it("AsyncValidation(ValidValidation) or ValidValidation should return AsyncValidation(ValidValidation)", async () => {
+            const res = createAsyncValidation
+                .validValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
                 .or(
-                    invalidValidation([
-                        { message: "invalid", value: "invalid value" },
+                    createValidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: true,
+                            inverted: false,
+                        },
                     ])
                 );
             expect(res.fork).toBeDefined();
@@ -315,21 +827,24 @@ describe("Validation", () => {
             expect(await res.getResult()).toBeUndefined();
         });
 
-        it("asyncValidation(Valid) or asyncValidation(Valid) return asyncValidation(Valid)", async () => {
-            const res = asyncValidation
-                .validValidation()
-                .or(asyncValidation.validValidation());
-            expect(res.fork).toBeDefined();
-
-            expect(await res.getResult()).toBeUndefined();
-        });
-
-        it("asyncValidation(Valid) or asyncValidation(InValid) should return asyncValidation(Valid)", async () => {
-            const res = asyncValidation
-                .validValidation()
+        it("AsyncValidation(ValidValidation) or InvalidValidation should return AsyncValidation(ValidValidation)", async () => {
+            const res = createAsyncValidation
+                .validValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
                 .or(
-                    asyncValidation.invalidValidation([
-                        { message: "invalid", value: "invalid value" },
+                    createInvalidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: false,
+                            inverted: false,
+                        },
                     ])
                 );
             expect(res.fork).toBeDefined();
@@ -337,61 +852,771 @@ describe("Validation", () => {
             expect(await res.getResult()).toBeUndefined();
         });
 
-        it("asyncValidation(Invalid) or Valid should return asyncValidation(Value)", async () => {
-            const res = asyncValidation
-                .invalidValidation([
-                    { message: "invalid", value: "invalid value" },
+        it("AsyncValidation(ValidValidation) or AsyncValidation(ValidValidation) should return an AsyncValidation(ValidValidation)", async () => {
+            const res = createAsyncValidation
+                .validValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: true,
+                        inverted: false,
+                    },
                 ])
-                .or(validValidation());
+                .or(
+                    createAsyncValidation.validValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: true,
+                            inverted: false,
+                        },
+                    ])
+                );
             expect(res.fork).toBeDefined();
 
             expect(await res.getResult()).toBeUndefined();
         });
 
-        it("asyncValidation(Invalid) or Invalid should return asyncValidation(Invalid)", async () => {
-            const res = asyncValidation
-                .invalidValidation([
-                    { message: "invalid1", value: "invalid value" },
+        it("AsyncValidation(ValidValidation) or AsyncValidation(InvalidValidation) should return an AsyncValidation(ValidValidation)", async () => {
+            const res = createAsyncValidation
+                .validValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: true,
+                        inverted: false,
+                    },
                 ])
                 .or(
-                    invalidValidation([
-                        { message: "invalid2", value: "invalid value" },
+                    createAsyncValidation.invalidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: false,
+                            inverted: false,
+                        },
+                    ])
+                );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toBeUndefined();
+        });
+
+        it("AsyncValidation(InvalidValidation) or ValidValidation should return AsyncValidation(ValidValidation)", async () => {
+            const res = createAsyncValidation
+                .invalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: false,
+                        inverted: false,
+                    },
+                ])
+                .or(
+                    createValidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: true,
+                            inverted: false,
+                        },
+                    ])
+                );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toBeUndefined();
+        });
+
+        it("AsyncValidation(InvalidValidation) or InvalidValidation should return AsyncValidation(InvalidValidation)", async () => {
+            const res = createAsyncValidation
+                .invalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: false,
+                        inverted: false,
+                    },
+                ])
+                .or(
+                    createInvalidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: false,
+                            inverted: false,
+                        },
                     ])
                 );
             expect(res.fork).toBeDefined();
 
             expect(await res.getResult()).toEqual([
-                { message: "invalid1", value: "invalid value" },
-                { message: "invalid2", value: "invalid value" },
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
             ]);
         });
 
-        it("asyncValidation(Invalid) or asyncValidation(Valid) return asyncValidation(Valid)", async () => {
-            const res = asyncValidation
+        it("AsyncValidation(InvalidValidation) or AsyncValidation(ValidValidation) should return an AsyncValidation(ValidValidation)", async () => {
+            const res = createAsyncValidation
                 .invalidValidation([
-                    { message: "invalid", value: "invalid value" },
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: false,
+                        inverted: false,
+                    },
                 ])
-                .or(asyncValidation.validValidation());
+                .or(
+                    createAsyncValidation.validValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: true,
+                            inverted: false,
+                        },
+                    ])
+                );
             expect(res.fork).toBeDefined();
 
             expect(await res.getResult()).toBeUndefined();
         });
 
-        it("asyncValidation(Invalid) or asyncValidation(InValid) should return asyncValidation(Valid)", async () => {
-            const res = asyncValidation
+        it("AsyncValidation(InvalidValidation) or AsyncValidation(InValidValidation) should return an AsyncValidation(InvalidValidation)", async () => {
+            const res = createAsyncValidation
                 .invalidValidation([
-                    { message: "invalid1", value: "invalid value" },
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: false,
+                        inverted: false,
+                    },
                 ])
                 .or(
-                    asyncValidation.invalidValidation([
-                        { message: "invalid2", value: "invalid value" },
+                    createAsyncValidation.invalidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: false,
+                            inverted: false,
+                        },
                     ])
                 );
             expect(res.fork).toBeDefined();
 
             expect(await res.getResult()).toEqual([
-                { message: "invalid1", value: "invalid value" },
-                { message: "invalid2", value: "invalid value" },
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
+            ]);
+        });
+    });
+
+    describe("not", () => {
+        it("should convert ValidValidation to InvalidValidation", () => {
+            const res = createValidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
+            ]).not();
+
+            expect(res.isValid).toBe(false);
+            expect(res.getResult()).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: true,
+                },
+            ]);
+            expect(res.results).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: true,
+                },
+            ]);
+        });
+
+        it("should convert InvalidValidation to ValidValidation", () => {
+            const res = createInvalidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+            ]).not();
+
+            expect(res.isValid).toBe(true);
+            expect(res.getResult()).toBeUndefined();
+            expect(res.results).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: true,
+                },
+            ]);
+        });
+    });
+
+    describe("xor", () => {
+        it("ValidValidation xor ValidValidation should return InvalidValidation", () => {
+            const res = createValidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
+            ]).xor(
+                createValidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+            );
+            expect(res.isValid).toBe(false);
+            expect(res.getResult()).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: true,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: true,
+                    inverted: true,
+                },
+            ]);
+            expect(res.results).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: true,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: true,
+                    inverted: true,
+                },
+            ]);
+        });
+        it("ValidValidation xor inValidValidation should return ValidValidation", () => {
+            const res = createValidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
+            ]).xor(
+                createInvalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: false,
+                        inverted: false,
+                    },
+                ])
+            );
+            expect(res.isValid).toBe(true);
+            expect(res.getResult()).toBeUndefined();
+            expect(res.results).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
+            ]);
+        });
+
+        it("ValidValidation xor AsyncValidation(ValidValidation) should return an AsyncValidation(InvalidValidation)", async () => {
+            const res = createValidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
+            ]).xor(
+                createAsyncValidation.validValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+            );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: true,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: true,
+                    inverted: true,
+                },
+            ]);
+        });
+
+        it("ValidValidation or AsyncValidation(InValidValidation) should return an AsyncValidation(ValidValidation)", async () => {
+            const res = createValidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: false,
+                },
+            ]).or(
+                createAsyncValidation.invalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: false,
+                        inverted: false,
+                    },
+                ])
+            );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toBeUndefined();
+        });
+        it("InvalidValidation xor ValidValidation should return ValidValidation", () => {
+            const res = createInvalidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+            ]).xor(
+                createValidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+            );
+            expect(res.isValid).toBe(true);
+            expect(res.getResult()).toBeUndefined();
+            expect(res.results).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: true,
+                    inverted: false,
+                },
+            ]);
+        });
+        it("InvalidValidation xor InvalidValidation should return InvalidValidation", () => {
+            const res = createInvalidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+            ]).xor(
+                createInvalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: false,
+                        inverted: false,
+                    },
+                ])
+            );
+            expect(res.isValid).toBe(false);
+            expect(res.getResult()).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
+            ]);
+            expect(res.results).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
+            ]);
+        });
+
+        it("InvalidValidation xor AsyncValidation(ValidValidation) should return an AsyncValidation(ValidValidation)", async () => {
+            const res = createInvalidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+            ]).xor(
+                createAsyncValidation.validValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+            );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toBeUndefined();
+        });
+
+        it("InvalidValidation xor AsyncValidation(InValidValidation) should return an AsyncValidation(InvalidValidation)", async () => {
+            const res = createInvalidValidation([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+            ]).xor(
+                createAsyncValidation.invalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isBar",
+                        valid: false,
+                        inverted: false,
+                    },
+                ])
+            );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
+            ]);
+        });
+
+        it("AsyncValidation(ValidValidation) xor ValidValidation should return AsyncValidation(InvalidValidation)", async () => {
+            const res = createAsyncValidation
+                .validValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+                .xor(
+                    createValidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: true,
+                            inverted: false,
+                        },
+                    ])
+                );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: true,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: true,
+                    inverted: true,
+                },
+            ]);
+        });
+
+        it("AsyncValidation(ValidValidation) xor InvalidValidation should return AsyncValidation(ValidValidation)", async () => {
+            const res = createAsyncValidation
+                .validValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+                .xor(
+                    createInvalidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: false,
+                            inverted: false,
+                        },
+                    ])
+                );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toBeUndefined();
+        });
+
+        it("AsyncValidation(ValidValidation) xor AsyncValidation(ValidValidation) should return an AsyncValidation(InvalidValidation)", async () => {
+            const res = createAsyncValidation
+                .validValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+                .xor(
+                    createAsyncValidation.validValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: true,
+                            inverted: false,
+                        },
+                    ])
+                );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: true,
+                    inverted: true,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: true,
+                    inverted: true,
+                },
+            ]);
+        });
+
+        it("AsyncValidation(ValidValidation) or AsyncValidation(InvalidValidation) should return an AsyncValidation(ValidValidation)", async () => {
+            const res = createAsyncValidation
+                .validValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: true,
+                        inverted: false,
+                    },
+                ])
+                .xor(
+                    createAsyncValidation.invalidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: false,
+                            inverted: false,
+                        },
+                    ])
+                );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toBeUndefined();
+        });
+
+        it("AsyncValidation(InvalidValidation) xor ValidValidation should return AsyncValidation(ValidValidation)", async () => {
+            const res = createAsyncValidation
+                .invalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: false,
+                        inverted: false,
+                    },
+                ])
+                .xor(
+                    createValidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: true,
+                            inverted: false,
+                        },
+                    ])
+                );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toBeUndefined();
+        });
+
+        it("AsyncValidation(InvalidValidation) xor InvalidValidation should return AsyncValidation(InvalidValidation)", async () => {
+            const res = createAsyncValidation
+                .invalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: false,
+                        inverted: false,
+                    },
+                ])
+                .xor(
+                    createInvalidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: false,
+                            inverted: false,
+                        },
+                    ])
+                );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
+            ]);
+        });
+
+        it("AsyncValidation(InvalidValidation) xor AsyncValidation(ValidValidation) should return an AsyncValidation(ValidValidation)", async () => {
+            const res = createAsyncValidation
+                .invalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: false,
+                        inverted: false,
+                    },
+                ])
+                .xor(
+                    createAsyncValidation.validValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: true,
+                            inverted: false,
+                        },
+                    ])
+                );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toBeUndefined();
+        });
+
+        it("AsyncValidation(InvalidValidation) xor AsyncValidation(InValidValidation) should return an AsyncValidation(InvalidValidation)", async () => {
+            const res = createAsyncValidation
+                .invalidValidation([
+                    {
+                        value: "value",
+                        predicate: "isFoo",
+                        valid: false,
+                        inverted: false,
+                    },
+                ])
+                .xor(
+                    createAsyncValidation.invalidValidation([
+                        {
+                            value: "value",
+                            predicate: "isBar",
+                            valid: false,
+                            inverted: false,
+                        },
+                    ])
+                );
+            expect(res.fork).toBeDefined();
+
+            expect(await res.getResult()).toEqual([
+                {
+                    value: "value",
+                    predicate: "isFoo",
+                    valid: false,
+                    inverted: false,
+                },
+                {
+                    value: "value",
+                    predicate: "isBar",
+                    valid: false,
+                    inverted: false,
+                },
             ]);
         });
     });
