+++ src/validators/array.ts
@@ -2,8 +2,9 @@ import { validator, Validator, getEntry } from "../Validator";
 import { SyncValidation, AsyncValidation } from "../Validation";
 import { and, addKeyToMessage } from "../utils";
 
-export const isArray = validator((value: any) =>
-    Array.isArray(value) ? undefined : "value must be an array"
+export const isArray = validator(
+    (value: any) => Array.isArray(value),
+    "value must be an array"
 );
 
 export const arrayOf = <T extends SyncValidation | AsyncValidation>(
