+++ src/validators/length.ts
@@ -1,36 +1,31 @@
 import { validator } from "../Validator";
 
 export const hasLengthOf = (x: number) =>
-    validator((value) =>
-        value && value.length === x
-            ? undefined
-            : `value must have a length of ${x}`
+    validator(
+        (value) => value && value.length === x,
+        `value must have a length of ${x}`
     );
 
 export const hasLengthGt = (min: number) =>
-    validator((value) =>
-        value && value.length > min
-            ? undefined
-            : `value must have a length greater than ${min}`
+    validator(
+        (value) => (value && value.length > min ? true : false),
+        `value must have a length greater than ${min}`
     );
 
 export const hasLengthGte = (min: number) =>
-    validator((value) =>
-        value && value.length >= min
-            ? undefined
-            : `value must have a length of at least ${min}`
+    validator(
+        (value) => value && value.length >= min,
+        `value must have a length of at least ${min}`
     );
 
 export const hasLengthLt = (max: number) =>
-    validator((value) =>
-        value && value.length < max
-            ? undefined
-            : `value must have a length less than ${max}`
+    validator(
+        (value) => value && value.length < max,
+        `value must have a length less than ${max}`
     );
 
 export const hasLengthLte = (max: number) =>
-    validator((value) =>
-        value && value.length <= max
-            ? undefined
-            : `value must have a length of at most ${max}`
+    validator(
+        (value) => value && value.length <= max,
+        `value must have a length of at most ${max}`
     );
