+++ src/validators/object.ts
@@ -2,8 +2,9 @@ import { addKeyToMessage, and } from "../utils";
 import { validator, Validator } from "../Validator";
 import { SyncValidation, AsyncValidation } from "../Validation";
 
-export const isObject = validator((value) =>
-    typeof value === "object" ? undefined : "value must be an object"
+export const isObject = validator(
+    (value) => typeof value === "object",
+    "value must be an object"
 );
 
 export const hasNoExtraneousKeys = (keys: string[]) =>
@@ -15,10 +16,8 @@ export const hasNoExtraneousKeys = (keys: string[]) =>
             (key) => !keys.includes(key)
         );
 
-        return extraneousKeys.length > 0
-            ? `Value has extraneous keys: ${extraneousKeys.join(", ")}`
-            : undefined;
-    });
+        return extraneousKeys.length <= 0;
+    }, `Value has extraneous keys`);
 
 export const isExactObject = (keys: string[]) =>
     isObject.and(hasNoExtraneousKeys(keys));
