+++ src/validators/string.ts
@@ -1,10 +1,12 @@
 import { validator } from "../Validator";
 
-export const isString = validator((value) =>
-    typeof value === "string" ? undefined : "value must be a string"
+export const isString = validator(
+    (value) => typeof value === "string",
+    "value must be a string"
 );
 
 export const match = (pattern: RegExp) =>
-    validator((value) =>
-        pattern.test(value) ? undefined : `value must match pattern ${pattern}`
+    validator(
+        (value) => pattern.test(value),
+        `value must match pattern ${pattern}`
     );
